一大波修改

+++ src/main/java/com/blog/api/repo/SysUserRepository.java
@@ -6,9 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.io.Serializable;
 import java.util.stream.Stream;
 @Repository
-public interface SysUserRepository extends BaseRepository<SysUser, Long> {
+public interface SysUserRepository extends BaseRepository<SysUser, Integer> {
 
     //接口方法的名称，符合约定则无需实现即可访问
     @Query("select u from SysUser u ")
@@ -18,4 +19,7 @@ public interface SysUserRepository extends BaseRepository<SysUser, Long> {
 
     @Query(value = "select u from SysUser u where account=?1")
     SysUser getUser(String account);
+
+
+
 }
